refactoring; password reset with soft delete

+++ app/Http/Controllers/Authentication/PasswordResetController.php
@@ -6,7 +6,6 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
 use Exception;
 use Illuminate\Support\Facades\Hash;
-// Custom
 use App\Models\PasswordResets\PasswordReset;
 use App\Notifications\Auth\ChangePasswordNotification;
 use App\Http\Requests\Auth\ForgotPassword\UpdatePasswordRequest;
@@ -22,18 +21,22 @@ class PasswordResetController extends Controller
     public function __invoke(UpdatePasswordRequest $request)
     {
         try {
-            $token = $this->model->where("token", $request->token)->first();
 
-            if (!$token) {
-                throw new Exception("Token inválido");
-            }
+            DB::transaction(function () use ($request) {
 
-            $token->user->update([
-                "password" => Hash::make($request->password)
-            ]);
+                $token = $this->model->where("token", $request->token)->first();
 
-            $token->delete();
-            $token->user->notify(new ChangePasswordNotification($token->user));
+                if (!$token || $token->trashed()) {
+                    throw new Exception("Token inválido");
+                }
+
+                $token->user->update([
+                    "password" => Hash::make($request->password)
+                ]);
+
+                $token->delete();
+                $token->user->notify(new ChangePasswordNotification($token->user));
+            });
 
             return response(["message" => "Senha alterada com sucesso!"], 200);
         } catch (Exception $e) {
